caproto-backend/requirements.txt
aiohttp
aiohttp_auth
aiohttp_session[aioredis]
aiohttp-swagger[performance]
asynctest
fixtures
uvloop
caproto
pyjwt
